TheGuard/requirements.txt
# Web Framework
Flask>=2.0.0

# Network Analysis
scapy>=2.4.5
pyshark>=0.4.5
python-snort>=0.3.0

# Machine Learning & Analysis
tensorflow>=2.0.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0

# Utilities
PyYAML>=5.4.0
requests>=2.26.0
watchdog>=2.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Security
cryptography>=3.4.0
bcrypt>=3.2.0

# Documentation
Sphinx>=4.5.0